Выделение типа колонки в отдельную абстракцию
