+++ tests/Queries/ColumnTest.php
@@ -53,434 +53,6 @@ class ColumnTest extends TestCase
     }
 
     /**
-     * integer
-     */
-    public function testInteger(): void
-    {
-        $column = Column::create()->name('column1')->integer();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'integer',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * bigInteger
-     */
-    public function testBigInteger(): void
-    {
-        $column = Column::create()->name('column1')->bigInteger();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'bigInteger',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * binary
-     */
-    public function testBinary(): void
-    {
-        $column = Column::create()->name('column1')->binary();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'binary',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * boolean
-     */
-    public function testBoolean(): void
-    {
-        $column = Column::create()->name('column1')->boolean();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'boolean',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * char
-     */
-    public function testChar(): void
-    {
-        $column = Column::create()->name('column1')->char();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'char',
-            'params' => ['length' => 255],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * date
-     */
-    public function testDate(): void
-    {
-        $column = Column::create()->name('column1')->date();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'date',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * dateTime
-     */
-    public function testDateTime(): void
-    {
-        $column = Column::create()->name('column1')->dateTime();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'dateTime',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * decimal
-     */
-    public function testDecimal(): void
-    {
-        $column = Column::create()->name('column1')->decimal();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'decimal',
-            'params' => [
-                'unsigned' => false,
-                'total' => 8,
-                'places' => 2,
-            ],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * double
-     */
-    public function testDouble(): void
-    {
-        $column = Column::create()->name('column1')->double();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'double',
-            'params' => [
-                'unsigned' => false,
-                'total' => null,
-                'places' => null,
-            ],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * enum
-     */
-    public function testEnum(): void
-    {
-        $column = Column::create()->name('column1')->enum(['enum1', 'enum2']);
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'enum',
-            'params' => ['enums' => ['enum1', 'enum2'],],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * float
-     */
-    public function testFloat(): void
-    {
-        $column = Column::create()->name('column1')->float();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'float',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * json
-     */
-    public function testJson(): void
-    {
-        $column = Column::create()->name('column1')->json();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'json',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * longText
-     */
-    public function testLongText(): void
-    {
-        $column = Column::create()->name('column1')->longText();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'longText',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * mediumInteger
-     */
-    public function testMediumInteger(): void
-    {
-        $column = Column::create()->name('column1')->mediumInteger();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'mediumInteger',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * mediumText
-     */
-    public function testMediumText(): void
-    {
-        $column = Column::create()->name('column1')->mediumText();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'mediumText',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * smallInteger
-     */
-    public function testSmallInteger(): void
-    {
-        $column = Column::create()->name('column1')->smallInteger();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'smallInteger',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * tinyInteger
-     */
-    public function testTinyInteger(): void
-    {
-        $column = Column::create()->name('column1')->tinyInteger();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'tinyInteger',
-            'params' => ['unsigned' => false],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * string
-     */
-    public function testString(): void
-    {
-        $column = Column::create()->name('column1')->string();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'string',
-            'params' => ['length' => 255],
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * text
-     */
-    public function testText(): void
-    {
-        $column = Column::create()->name('column1')->text();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'text',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * time
-     */
-    public function testTime(): void
-    {
-        $column = Column::create()->name('column1')->time();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'time',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * timestamp
-     */
-    public function testTimestamp(): void
-    {
-        $column = Column::create()->name('column1')->timestamp();
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'timestamp',
-            'params' => null,
-            'nullable' => false,
-            'default' => null,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
      * nullable
      */
     public function testNullable(): void
@@ -512,26 +84,6 @@ class ColumnTest extends TestCase
             'params' => null,
             'nullable' => false,
             'default' => 100,
-            'unique' => null,
-            'primary' => null,
-            'index' => null,
-            'foreign' => null,
-            'rename' => null,
-        ], $column->getStructure());
-    }
-
-    /**
-     * unsigned
-     */
-    public function testUnsigned(): void
-    {
-        $column = Column::create()->name('column1')->integer(true);
-        $this->assertEquals([
-            'columnName' => 'column1',
-            'type' => 'integer',
-            'params' => ['unsigned' => true],
-            'nullable' => false,
-            'default' => null,
             'unique' => null,
             'primary' => null,
             'index' => null,
